добавил static_cast<int> для аптечек

+++ MyStrategy.cpp
@@ -677,7 +677,8 @@ void getAttackingData2(
 				}
 			}
 			if (isGot) continue;
-			if ((lb.position.x - enemy.position.x) * (me.position.x - enemy.position.x) < 0) continue;
+			if ((static_cast<int>(lb.position.x) - static_cast<int>(enemy.position.x)) * 
+				(static_cast<int>(me.position.x) - enemy.position.x) < 0) continue;
 
 			const auto dist = MathHelper::getMHDist(me.position, lb.position);
 			if (dist < minMHDist)
